Move parsers for remaining function-based formats
This covers all of the legacy and CSSC4 functional formats for HSV, HWB,
Lab, LCh, and CMYK colors, but not the CSSC4 color space formats based
on the `color()` function.
