+++ src/cmyk.rs
@@ -1,6 +1,18 @@
 use std::fmt;
 
-use crate::{rgb::RGBA, types::Scalar, Color, Format};
+use nom::{
+    bytes::complete::tag_no_case,
+    character::complete::{char, space0, space1},
+    combinator::{all_consuming, opt},
+    IResult,
+};
+
+use crate::{
+    parser::{modern_alpha, number_or_percentage},
+    rgb::RGBA,
+    types::Scalar,
+    Color, Format,
+};
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct CMYK {
@@ -81,6 +93,35 @@ impl CMYK {
     }
 }
 
+pub(crate) fn parse_cmyk_color(input: &str) -> IResult<&str, Color> {
+    all_consuming(parse_css_device_cmyk)(input.trim())
+}
+
+// Parse CMYK colors as the `device-cmyk()` function defined in CSS Color 5.  The simpler `cmyk()`
+// is also accepted.  We have no color profile info here, so all CMYK colors are represented as
+// uncalibrated colors using the naive RGB conversion.
+
+fn parse_css_device_cmyk(input: &str) -> IResult<&str, Color> {
+    let (input, _) = opt(tag_no_case("device-"))(input)?;
+    let (input, _) = tag_no_case("cmyk(")(input)?;
+    let (input, _) = space0(input)?;
+    let (input, c) = number_or_percentage(input, 1.0)?;
+    let (input, _) = space1(input)?;
+    let (input, m) = number_or_percentage(input, 1.0)?;
+    let (input, _) = space1(input)?;
+    let (input, y) = number_or_percentage(input, 1.0)?;
+    let (input, _) = space1(input)?;
+    let (input, k) = number_or_percentage(input, 1.0)?;
+    // accept alpha component for compatibility, but not currently supported for CMYK
+    let (input, _alpha) = modern_alpha(input)?;
+    let (input, _) = space0(input)?;
+    let (input, _) = char(')')(input)?;
+
+    let c = Color::from_cmyk(c, m, y, k);
+
+    Ok((input, c))
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -121,5 +162,78 @@ mod tests {
 
         let c3 = CMYK::new(0.25, 0.1, 0.0, 0.5);
         assert_eq!("cmyk(25, 10, 0, 50)", c3.to_color_string(Format::Spaces));
+    }
+
+    fn parse_color(input: &str) -> Option<Color> {
+        parse_cmyk_color(input).ok().map(|(_, c)| c)
+    }
+
+    #[test]
+    fn parse_css_device_cmyk_syntax() {
+        assert_eq!(
+            Some(Color::from_cmyk(0.8, 0.2, 0.6, 0.4)),
+            parse_color("device-cmyk(80% 20% 60% 40%)")
+        );
+        assert_eq!(
+            Some(Color::from_cmyk(0.8, 0.2, 0.6, 0.4)),
+            parse_color("device-cmyk(0.8 0.2 0.6 0.4)")
+        );
+        assert_eq!(
+            Some(Color::from_cmyk(0.8, 0.2, 0.6, 0.4)),
+            parse_color("device-cmyk(80% 0.2 0.6 40%)")
+        );
+
+        assert_eq!(
+            Some(Color::from_cmyk(0.8, 0.2, 0.6, 0.4)),
+            parse_color("device-cmyk(80.0% 20.000% 60% 40.%)")
+        );
+        assert_eq!(
+            Some(Color::from_cmyk(0.8, 0.2, 0.6, 0.4)),
+            parse_color("device-cmyk(0.800 0.2 .6 0.40)")
+        );
+
+        assert_eq!(
+            Some(Color::red()),
+            parse_color("device-cmyk(0% 100% 100% 0%)")
+        );
+        assert_eq!(
+            Some(Color::green()),
+            parse_color("device-cmyk(100% 0% 100% 50%)")
+        );
+        assert_eq!(
+            Some(Color::blue()),
+            parse_color("device-cmyk(100% 100% 0% 0%)")
+        );
+        assert_eq!(Some(Color::yellow()), parse_color("device-cmyk(0 0 1 0)"));
+        assert_eq!(
+            Some(Color::from_rgb(255, 165, 0)), // orange
+            parse_color("device-cmyk(0 0.353 1 0)")
+        );
+        assert_eq!(Some(Color::purple()), parse_color("device-cmyk(0 1 0 0.5)"));
+
+        // `cmyk` is equivalent to `device-cmyk`
+        assert_eq!(
+            Some(Color::from_cmyk(0.8, 0.2, 0.6, 0.4)),
+            parse_color("cmyk(0.8 0.2 0.6 0.4)")
+        );
+
+        // function names are case-insensitive
+        assert_eq!(
+            Some(Color::from_cmyk(0.8, 0.2, 0.6, 0.4)),
+            parse_color("Device-CMYK(80% 20% 60% 40%)")
+        );
+
+        // alpha value is allowed for compatibility, but is ignored because it
+        // isn't currently supported by the color library
+        assert_eq!(
+            Some(Color::from_cmyk(0.8, 0.2, 0.6, 0.4)),
+            parse_color("device-cmyk(80% 20% 60% 40% / 0.5)")
+        );
+
+        assert_eq!(None, parse_color("device-cmyk(0,1,1,0)"));
+        assert_eq!(None, parse_color("device-cmyk(0 1 1)"));
+        assert_eq!(None, parse_color("device-cmyk(0 1)"));
+        assert_eq!(None, parse_color("device-cmyk(50%)"));
+        assert_eq!(None, parse_color("device-cmyk(0 1 0.5 1 0)"));
     }
 }
